refactor: separate logic to controller and model
