+++ src/models/User.js
@@ -1,0 +1,16 @@
+import bcrypt from 'bcrypt';
+
+export default class User {
+  static async create({
+    first_name, last_name, email, password,
+  } = {}) {
+    const hashedPassword = await bcrypt.hash(password, 10);
+    const {
+      rows: [record],
+    } = await db.query(
+      'INSERT INTO "user"(first_name, last_name, email, password) VALUES ($1, $2, $3, $4) RETURNING *;',
+      [first_name, last_name, email, hashedPassword],
+    );
+    return record;
+  }
+}
